Implemented strict c-string formatting

+++ src/ti/access.c
@@ -129,7 +129,7 @@ int ti_access_check_err(
                 user->name->n, (char *) user->name->data,
                 ti_auth_mask_to_str(mask),
                 prefix,
-                name_sz, name);
+                (int) name_sz, name);
     }
     return e->nr;
 }
@@ -161,11 +161,11 @@ int ti_access_more_check(
         ex_set(e, EX_FORBIDDEN,
                 "user `%.*s` has privileges (`%s`) that user `%.*s` is "
                 "missing on scope `%s%.*s`",
-                n->name->n, (char *) n->name->data,
+                (int) n->name->n, (char *) n->name->data,
                 ti_auth_mask_to_str(nm),
                 o->name->n, (char *) o->name->data,
                 prefix,
-                name_sz, name);
+                (int) name_sz, name);
     }
     return e->nr;
 }
@@ -191,10 +191,10 @@ int ti_access_check_or_err(
         ex_set(e, EX_FORBIDDEN,
                 "user `%.*s` is missing the required privileges (`%s`) "
                 "on scope `%s%.*s`"DOC_GRANT,
-                user->name->n, (char *) user->name->data,
+                (int) user->name->n, (char *) user->name->data,
                 ti_auth_mask_to_str(mask),
                 prefix,
-                name_sz, name);
+                (int) name_sz, name);
     }
     return e->nr;
 }
